Redesigned Admin CPanel UI

+++ app/Http/Controllers/AdminUsersController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\User;
+use Illuminate\Support\Facades\Auth;
 
 class AdminUsersController extends Controller
 {
@@ -20,9 +21,16 @@ class AdminUsersController extends Controller
 
     public function destroy(User $user)
     {
-        if($user->delete()){
-            return redirect(route('admin.users.index'))->with('success', 'User deleted successfully');
+        if(Auth::user()->email != $user->email) {
+            if ($user->role->name != 'administrator') {
+                if ($user->delete()) {
+                    return back()->with('success', 'User deleted successfully');
+                }
+                return back()->with('failed', 'Failed in doing so.');
+            }
+            return back()->with('failed', 'Administrators cannot be deleted, promoted or demoted.');
         }
+        return back()->with('failed', 'You cannot delete logged in account.');
     }
 
     public function promote(User $user){
@@ -33,9 +41,10 @@ class AdminUsersController extends Controller
                     'role_id' => $promoted_id
                 ])
             ){
-                return redirect(route('admin.users.index'))->with('success', 'User promoted successfully');
+                return back()->with('success', 'User promoted successfully');
             }
         }
+        return back()->with('failed', 'This user cannot be promoted');
     }
     public function demote(User $user){
         if($user->role_id < 3 && $user->role_id > 1){
@@ -45,8 +54,9 @@ class AdminUsersController extends Controller
                     'role_id' => $demoted_id
                 ])
             ){
-                return redirect(route('admin.users.index'))->with('success', 'User demoted successfully');
+                return back()->with('success', 'User demoted successfully');
             }
         }
+        return back()->with('failed', 'This user cannot be demoted');
     }
 }
